SMB: Merge SMBPropagationClient into SMBExploitClient
