+++ monkey/agent_plugins/exploiters/smb/src/plugin.py
@@ -16,7 +16,6 @@ from infection_monkey.propagation_credentials_repository import IPropagationCred
 from .smb_exploit_client import SMBExploitClient
 from .smb_exploiter import SMBExploiter
 from .smb_options import SMBOptions
-from .smb_propagation_client import SMBPropagationClient
 
 logger = logging.getLogger(__name__)
 
@@ -33,11 +32,9 @@ class Plugin:
         **kwargs,
     ):
         smb_exploit_client = SMBExploitClient(agent_event_publisher)
-        smb_propagation_client = SMBPropagationClient(agent_event_publisher)
 
         self._smb_exploiter = SMBExploiter(
             smb_exploit_client,
-            smb_propagation_client,
             propagation_credentials_repository,
             agent_binary_repository,
         )
