+++ monkey/agent_plugins/exploiters/smb/src/smb_exploiter.py
@@ -12,7 +12,6 @@ from infection_monkey.utils.threading import interruptible_iter
 
 from .smb_exploit_client import SMBExploitClient
 from .smb_options import SMBOptions
-from .smb_propagation_client import SMBPropagationClient
 
 logger = logging.getLogger(__name__)
 
@@ -21,12 +20,10 @@ class SMBExploiter:
     def __init__(
         self,
         exploit_client: SMBExploitClient,
-        propagation_client: SMBPropagationClient,
         propagation_credentials_repository: IPropagationCredentialsRepository,
         agent_binary_repository: IAgentBinaryRepository,
     ):
         self._exploit_client = exploit_client
-        self._propagation_client = propagation_client
         self._propagation_credentials_repository = propagation_credentials_repository
         self._agent_binary_repository = agent_binary_repository
 
@@ -81,14 +78,14 @@ class SMBExploiter:
     ) -> bool:
         destination_path = get_agent_dst_path(host)
         agent_binary = self._agent_binary_repository.get_agent_binary(host.operating_system)
-        remote_path = self._propagation_client.copy_file(
+        remote_path = self._exploit_client.copy_file(
             agent_binary, host, destination_path, credentials, options
         )
         if not remote_path:
             logger.debug(f"Failed to copy agent to {host}")
             return False
 
-        if not self._propagation_client.run_agent(
+        if not self._exploit_client.run_agent(
             host, remote_path, destination_path, servers, current_depth, credentials, options
         ):
             logger.debug(f"Failed to run agent on {host}")
