+++ monkey/tests/unit_tests/agent_plugins/exploiters/smb/test_smb_exploiter.py
@@ -7,7 +7,6 @@ import pytest
 from agent_plugins.exploiters.smb.src.smb_exploit_client import SMBExploitClient
 from agent_plugins.exploiters.smb.src.smb_exploiter import SMBExploiter
 from agent_plugins.exploiters.smb.src.smb_options import SMBOptions
-from agent_plugins.exploiters.smb.src.smb_propagation_client import SMBPropagationClient
 from tests.data_for_tests.propagation_credentials import FULL_CREDENTIALS
 
 from common import OperatingSystem
@@ -23,12 +22,6 @@ CREDENTIALS: List[Credentials] = []
 def mock_exploit_client() -> SMBExploitClient:
     client = MagicMock(spec=SMBExploitClient)
     client.authenticate.return_value = True
-    return client
-
-
-@pytest.fixture
-def mock_propagation_client() -> SMBPropagationClient:
-    client = MagicMock(spec=SMBPropagationClient)
     client.copy_file.return_value = "path"
     client.run_agent.return_value = True
     return client
@@ -44,12 +37,10 @@ def mock_credentials_repository() -> IPropagationCredentialsRepository:
 @pytest.fixture
 def smb_exploiter(
     mock_exploit_client: SMBExploitClient,
-    mock_propagation_client: SMBPropagationClient,
     mock_credentials_repository: IPropagationCredentialsRepository,
 ) -> SMBExploiter:
     return SMBExploiter(
         mock_exploit_client,
-        mock_propagation_client,
         mock_credentials_repository,
         MagicMock(spec=IAgentBinaryRepository),
     )
@@ -101,9 +92,9 @@ def test_exploit_host__propagation_succeeds(
 
 def test_exploit_host__copy_fails(
     smb_exploiter: SMBExploiter,
-    mock_propagation_client: SMBPropagationClient,
+    mock_exploit_client: SMBExploitClient,
 ):
-    mock_propagation_client.copy_file.return_value = None
+    mock_exploit_client.copy_file.return_value = None
 
     result = run_smb_exploiter(smb_exploiter)
     assert result.exploitation_success
@@ -112,9 +103,9 @@ def test_exploit_host__copy_fails(
 
 def test_exploit_host__run_agent_fails(
     smb_exploiter: SMBExploiter,
-    mock_propagation_client: SMBPropagationClient,
+    mock_exploit_client: SMBExploitClient,
 ):
-    mock_propagation_client.run_agent.return_value = False
+    mock_exploit_client.run_agent.return_value = False
 
     result = run_smb_exploiter(smb_exploiter)
     assert result.exploitation_success
@@ -134,9 +125,9 @@ def test_exploit_host__exploit_fails_on_authentication_error(
 
 def test_exploit_host__propagation_fails_on_run_agent_error(
     smb_exploiter: SMBExploiter,
-    mock_propagation_client: SMBPropagationClient,
+    mock_exploit_client: SMBExploitClient,
 ):
-    mock_propagation_client.run_agent.side_effect = Exception()
+    mock_exploit_client.run_agent.side_effect = Exception()
 
     result = run_smb_exploiter(smb_exploiter)
     assert result.exploitation_success
